+++ src/index.js
@@ -2,6 +2,8 @@ import { Controller } from "stimulus";
 import { DirectUpload } from "@rails/activestorage";
 import Dropzone from "dropzone";
 
+const DEFAULT_PARALLEL_UPLOADS = 2;
+
 Dropzone.autoDiscover = false;
 
 class DropzoneController extends Controller {
@@ -9,6 +11,8 @@ class DropzoneController extends Controller {
 
   connect() {
     if (this.dropzone === undefined) {
+      this.queue = [];
+      this.uploadsInProgress = 0;
       this.dropzone = new Dropzone(this.element, {
         url: this.url,
         headers: this.headers,
@@ -16,7 +20,7 @@ class DropzoneController extends Controller {
         maxFiles: this.maxFiles,
         maxFilesize: this.maxFileSize,
         autoQueue: false,
-        parallelUploads: 2
+        parallelUploads: this.parallelUploads
       });
 
       this.bindEvents();
@@ -50,10 +54,16 @@ class DropzoneController extends Controller {
 
   handleFileAdded(file) {
     if (file.accepted) {
+      if (this.fileExists(file)) {
+        this.dropzone.removeFile(file);
+        return;
+      }
+
       this.dispatchEvent(this.fileAddedEvent, { detail: { file: file }});
 
       const controller = new DirectUploadController(this, file);
-      controller.start();
+      this.queue.push(controller);
+      this.runUploadQueue();
     }
   }
 
@@ -63,6 +73,8 @@ class DropzoneController extends Controller {
 
   handleFileSuccess(file) {
     this.dispatchEvent(this.fileSuccessEvent, { detail: { file: file }});
+    this.uploadsInProgress--;
+    this.runUploadQueue();
   }
 
   handleFileProgress(file, progress) {
@@ -71,10 +83,17 @@ class DropzoneController extends Controller {
 
   handleFileRemoved(file) {
     if (file.controller) {
-      file.controller.xhr.abort();
+      if (file.controller.xhr) {
+        file.controller.xhr.abort();
+        this.uploadsInProgress--;
+      } else {
+        this.removeFromQueue(file);
+      }
+
       file.controller.removeHiddenInput();
 
       this.dispatchEvent(this.fileRemovedEvent, { detail: { file: file }});
+      this.runUploadQueue();
     }
   }
 
@@ -137,6 +156,26 @@ class DropzoneController extends Controller {
     }
   }
 
+  fileExists(theFile) {
+    const files = this.dropzone.files.filter(file => (file.name == theFile.name && file.size == theFile.size));
+    return files.length > 1;
+  }
+
+  removeFromQueue(file) {
+    this.queue = this.queue.filter(controller => (controller.file.upload.uuid !== file.upload.uuid));
+  }
+
+  runUploadQueue() {
+    this.queue.forEach((controller, index, queue) => {
+      if (this.uploadsInProgress < this.parallelUploads) {
+        this.dispatchEvent(this.fileStartEvent, { detail: { file: controller.file }});
+        controller.start();
+        queue.splice(index, 1);
+        this.uploadsInProgress++;
+      }
+    });
+  }
+
   get headers() {
     return {
       "X-CSRF-Token": document.querySelector('meta[name="csrf-token"]').content
@@ -167,6 +206,10 @@ class DropzoneController extends Controller {
     return this.data.get("file-drop-event");
   }
 
+  get fileStartEvent() {
+    return this.data.get("file-start-event");
+  }
+
   get fileSuccessEvent() {
     return this.data.get("file-success-event");
   }
@@ -189,6 +232,10 @@ class DropzoneController extends Controller {
 
   get fileDropOverId() {
     return this.data.get("file-drop-over-id") || null;
+  }
+
+  get parallelUploads() {
+    return this.data.get("parallel-uploads") || DEFAULT_PARALLEL_UPLOADS;
   }
 }