+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2021-02-15
+- Add `parallel-uploads`, `file-start-event`
+- Prevent duplicate files from being added
+
 ## [1.1.1] - 2021-02-12
 - Prevent duplicate binding